+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Make the `Decimal` classes globally available
 * Allow passing a custom context to `PathEvaluator#getValue(context)`
 * Fix `TaskService#initSchedules()` attempting to recreate already existing system schedules
+* Add `Schema#addComputedField()` method
 
 ## 1.3.19 (2023-10-18)